descriptor/src/androidTest/java/org/im97mori/ble/descriptor/ReportReferenceTest.java
package org.im97mori.ble.descriptor;

import android.bluetooth.BluetoothGattDescriptor;
import android.os.Parcel;

import org.junit.Test;

import static org.im97mori.ble.BLEConstants.BASE_UUID;
import static org.junit.Assert.assertArrayEquals;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

public class ReportReferenceTest {

    @Test
    public void test_constructor001() {
        //@formatter:off
        byte[] value = new byte[2];
        value[ 0] = 0x01;
        value[ 1] = 0x02;
        //@formatter:on

        BluetoothGattDescriptor bluetoothGattDescriptor = new BluetoothGattDescriptor(BASE_UUID, 0);
        bluetoothGattDescriptor.setValue(value);

        ReportReference result = new ReportReference(bluetoothGattDescriptor);
        assertEquals(0x01, result.getReportId());
        assertEquals(0x02, result.getReportType());
    }

    @Test
    public void test_constructor002() {
        //@formatter:off
        byte[] value = new byte[2];
        value[ 0] = 0x01;
        value[ 1] = (byte) ReportReference.REPORT_TYPE_INPUT_REPORT;
        //@formatter:on

        BluetoothGattDescriptor bluetoothGattDescriptor = new BluetoothGattDescriptor(BASE_UUID, 0);
        bluetoothGattDescriptor.setValue(value);

        ReportReference result = new ReportReference(bluetoothGattDescriptor);
        assertTrue(result.isReportTypeInputReport());
        assertFalse(result.isReportTypeOutputReport());
        assertFalse(result.isReportTypeFeatureReport());
    }

    @Test
    public void test_constructor003() {
        //@formatter:off
        byte[] value = new byte[2];
        value[ 0] = 0x01;
        value[ 1] = (byte) ReportReference.REPORT_TYPE_OUTPUT_REPORT;
        //@formatter:on

        BluetoothGattDescriptor bluetoothGattDescriptor = new BluetoothGattDescriptor(BASE_UUID, 0);
        bluetoothGattDescriptor.setValue(value);

        ReportReference result = new ReportReference(bluetoothGattDescriptor);
        assertFalse(result.isReportTypeInputReport());
        assertTrue(result.isReportTypeOutputReport());
        assertFalse(result.isReportTypeFeatureReport());
    }

    @Test
    public void test_constructor004() {
        //@formatter:off
        byte[] value = new byte[2];
        value[ 0] = 0x01;
        value[ 1] = (byte) ReportReference.REPORT_TYPE_FEATURE_REPORT;
        //@formatter:on

        BluetoothGattDescriptor bluetoothGattDescriptor = new BluetoothGattDescriptor(BASE_UUID, 0);
        bluetoothGattDescriptor.setValue(value);

        ReportReference result = new ReportReference(bluetoothGattDescriptor);
        assertFalse(result.isReportTypeInputReport());
        assertFalse(result.isReportTypeOutputReport());
        assertTrue(result.isReportTypeFeatureReport());
    }

    @Test
    public void test_parcelable001() {
        //@formatter:off
        byte[] value = new byte[2];
        value[ 0] = 0x01;
        value[ 1] = 0x02;
        //@formatter:on

        BluetoothGattDescriptor bluetoothGattDescriptor = new BluetoothGattDescriptor(BASE_UUID, 0);
        bluetoothGattDescriptor.setValue(value);

        ReportReference result1 = new ReportReference(bluetoothGattDescriptor);
        Parcel parcel = Parcel.obtain();
        result1.writeToParcel(parcel, 0);
        parcel.setDataPosition(0);
        ReportReference result2 = ReportReference.CREATOR.createFromParcel(parcel);

        assertEquals(result1.getReportId(), result2.getReportId());
        assertEquals(result1.getReportType(), result2.getReportType());
    }

    @Test
    public void test_parcelable002() {
        //@formatter:off
        byte[] value = new byte[2];
        value[ 0] = 0x01;
        value[ 1] = 0x02;
        //@formatter:on

        BluetoothGattDescriptor bluetoothGattDescriptor = new BluetoothGattDescriptor(BASE_UUID, 0);
        bluetoothGattDescriptor.setValue(value);

        ReportReference result1 = new ReportReference(bluetoothGattDescriptor);
        assertArrayEquals(value, result1.getBytes());
    }

    @Test
    public void test_parcelable003() {
        //@formatter:off
        byte[] value = new byte[2];
        value[ 0] = 0x01;
        value[ 1] = 0x02;
        //@formatter:on

        BluetoothGattDescriptor bluetoothGattDescriptor = new BluetoothGattDescriptor(BASE_UUID, 0);
        bluetoothGattDescriptor.setValue(value);

        ReportReference result1 = new ReportReference(bluetoothGattDescriptor);
        ReportReference result2 = ReportReference.CREATOR.createFromByteArray(value);
        assertArrayEquals(result1.getBytes(), result2.getBytes());
    }
}
